pyproject.toml
[project]
name = "pw-homework-module-12"
version = "0.1.0"
description = ""
authors = [
    {name = "RuslanaPashkutska",email = "ryslana.pashkytska@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
packages = [
    { include = "src" }
]
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
